refactor sqlite election error handling
